refactor(vo) : member 관련 vo들을 vo package로 이동
